fix(build): restore macOS clippy health after native-xstate merge
Remove unused GuestVa import in hvf_aarch64_engine.rs left over from the
merge's carrick_guest_mem re-export cleanup.
Add an accurate `# Safety` section to the off-x86 fail-closed
`enter_translated` stub in carrick-dsr-x86, documenting that it is
unreachable outside target_arch = "x86_64" and never touches its context.
Fix an E0505 in trap.rs's FakeStageCopyHarness::zero_copy_eligible test
helper (plan.first() borrow outliving a `for chunk in plan` move) by
iterating `&plan` instead of moving it.

Also needed to actually reach a clean `just clippy` on this Mac (the
above three were masked by a hard compile error until fixed):
- carrick-runtime's Cargo.toml gated the `carrick-dsr-x86` dependency to
  the FreeBSD/x86_64 native lane, but lib.rs unconditionally re-exports
  its portable `x86_dsr_profiler_layout` for cross-host diagnostic
  tooling (carrick-cli's debug_layout.rs) — made the dependency
  unconditional, same as carrick-dsr-aarch64.
- Six SyscallDispatcher/free-fn helpers (dispatch/mod.rs, mem.rs, proc.rs,
  sysv.rs) are only ever called from the FreeBSD/x86_64 native lane
  (native_freebsd.rs) and were flagged dead-code on macOS; cfg-gated their
  dead-code allowance to that target, mirroring the existing
  host_socket_buffer_size pattern (whose own cfg_attr was also corrected —
  its only caller is cfg(test)-gated, so it was dead in a plain lib build
  even on macOS).
- Allowed clippy::expect_used in native_darwin/dsr/oracle.rs, an
  entirely #[cfg(test)]-gated live-execution oracle/fixture module,
  matching the project's existing test-fixture convention (e.g.
  tests/live_bhyve_x86.rs).
- Removed 3 unnecessary `as u64` casts in dispatch/time.rs (rlim_t is
  already u64 on this target).
- Removed an unnecessary `mut` in tests/runtime_loop.rs
  (HostFsBackend::set_file_contents takes &self).
- Split the `DebugCommand` import out of commands.rs's shared `use`
  block into its own cfg(platform-linux/freebsd/netbsd) gate — the
  macOS `Commands::Debug` arm only forwards to run_debug and never
  names the type.

Verified: `just clippy` exits 0. `cargo test -p carrick-vmm-hvf
-p carrick-dsr-x86`: carrick-vmm-hvf 111/111 pass; carrick-dsr-x86
115/116 pass, with the pre-existing signal_xstate_roundtrip failure
(requires real x86_64 CPUID/XGETBV; this Mac is aarch64) unrelated to
this change.

+++ crates/carrick-runtime/src/dispatch/time.rs
@@ -943,14 +943,14 @@ fn raise_host_nofile_backing(guest_soft: u64) {
     if unsafe { libc::getrlimit(libc::RLIMIT_NOFILE, &mut rl) } != 0 {
         return;
     }
-    let cur_soft = rl.rlim_cur as u64;
+    let cur_soft = rl.rlim_cur;
     // The host hard limit ceilings any raise. A non-positive / INFINITY hard
     // value (rlim_max <= 0 when reinterpreted) means "no finite cap" → use the
     // desired target directly.
     let hard = rl.rlim_max;
     let desired = guest_soft.saturating_add(HOST_FD_HEADROOM);
-    let target = if hard > 0 && (hard as u64) < desired {
-        hard as u64
+    let target = if hard > 0 && hard < desired {
+        hard
     } else {
         desired
     };
